- Moved `optimizelyReadyReceiver` to from `DiscoveryFragment` `BaseFragment` - Added `optimizelyReady` to `FragmentViewModel`   - Not sure if I need to call `onCompleted` - Fixed bug where `EditorialActivity` opened twice
