+++ app/src/main/java/com/kickstarter/viewmodels/DiscoveryFragmentViewModel.java
@@ -63,9 +63,6 @@ public interface DiscoveryFragmentViewModel {
     /** Call for project pagination. */
     void nextPage();
 
-    /** Call when Optimizely is ready. */
-    void optimizelyReady();
-
     /** Call when params from Discovery Activity change. */
     void paramsFromActivity(final DiscoveryParams params);
 
@@ -340,7 +337,6 @@ public interface DiscoveryFragmentViewModel {
     private final PublishSubject<Boolean> discoveryOnboardingLoginToutClick = PublishSubject.create();
     private final PublishSubject<Editorial> editorialClicked = PublishSubject.create();
     private final PublishSubject<Void> nextPage = PublishSubject.create();
-    private final PublishSubject<Void> optimizelyReady = PublishSubject.create();
     private final PublishSubject<DiscoveryParams> paramsFromActivity = PublishSubject.create();
     private final PublishSubject<Project> projectCardClicked = PublishSubject.create();
     private final PublishSubject<Void> refresh = PublishSubject.create();
@@ -386,9 +382,6 @@ public interface DiscoveryFragmentViewModel {
     }
     @Override public void editorialViewHolderClicked(final @NonNull Editorial editorial) {
       this.editorialClicked.onNext(editorial);
-    }
-    @Override public void optimizelyReady() {
-      this.optimizelyReady.onNext(null);
     }
     @Override public void projectCardViewHolderClicked(final @NonNull Project project) {
       this.projectCardClicked.onNext(project);
